Minimize STL dependencies
