+++ include/nanobind/nanobind.h
@@ -32,8 +32,6 @@
 #include <typeinfo>
 #include <utility>
 #include <new>
-#include <optional>
-#include <variant>
 
 // Implementation. The nb_*.h files should only be included through nanobind.h
 #include "nb_python.h"
